Add error types for request failures
Add specific types to the configuration-related errors emitted by
aws-request, so user code can get more specific information about what
went wrong with the request.

fiddlerwoaroof is the original author of this fix.  I should have just
pulled it in, but I copy pasta'd it.  The original fix is here:
https://github.com/fiddlerwoaroof/aws-sdk-lisp/commit/9872d4e727d75730253ac9b60772e5ab8d8d147e