Fixed s3 warnings.
This closes https://github.com/smashedtoatoms/aws-sdk-cl/issues/7

There was an issue where the code generation was generating a struct
called `delete-objects` which was conflicting with the generated aws
function with the same symbol in services/s3.  I updated structs to
append `type` to all values.  `delete-objects` is now
`delete-objects-type`.  I rebuilt all of the services, and it looks like
the previous :use change didn't get rendered, so that change is in
here as well.